#52 Schedules grid much more mobile friendly
