+++ src/client/src/components/views/dashboard/schedules/DashboardSchedulesView.vue
@@ -3,92 +3,17 @@
     <h4>Create New Schedule</h4>
     <ScheduleForm
       ref="createForm"
-      :audiences="audiences"
-      :disabled="false" />
+      :audiences="audiences" />
     <Button
       class="input"
       label="Create"
       @click="handleCreateClick"
       :disabled="controlsDisabled" />
-    <DataTable
-      :loading="!isLoaded"
-      data-key="id"
-      :value="schedules"
-      :expanded-rows="expandedRows"
-      sort-field="start"
-      :sort-order="1">
-      <template #header>
-        <h4>Upcoming Schedules</h4>
-      </template>
-      <template>
-        <Column
-          expander
-          style="width: 5rem" />
-        <Column
-          field="audience.name"
-          header="Audience" />
-        <Column
-          field="start"
-          header="Date"
-          sortable>
-          <template #body="{ data }">
-            {{ parseDate(data.start).toLocaleDateString() }}
-          </template>
-        </Column>
-        <Column
-          field="end"
-          header="Time">
-          <template #body="{ data }">
-            {{ formatTimeslot(parseDate(data.start)) }} - {{ formatTimeslot(parseDate(data.end)) }}
-          </template>
-        </Column>
-        <Column class="grid-action-col">
-          <template #body="{ data }">
-            <Button
-              v-if="canEdit(data)"
-              icon="pi pi-pencil"
-              severity="secondary"
-              @click="handleEditScheduleActionClick(data)"
-              rounded
-              outlined />
-            <Button
-              v-if="data.timeslots.length === 0"
-              class="grid-action-col__item"
-              icon="pi pi-trash"
-              severity="danger"
-              @click="handleDeleteScheduleActionClick(data)"
-              rounded
-              outlined />
-          </template>
-        </Column>
-      </template>
-      <template #expansion="rowProps">
-        <DashboardSchedulesTimeslotsTable
-          :schedule="rowProps.data"
-          :performers="performers"
-          :disabled="isSubmitting || showEditScheduleDialog || showDeleteScheduleDialog"
-          @dialog-state-change="(value) => (isEditingTimeslot = value)" />
-      </template>
-    </DataTable>
-    <FormDialog
-      header="Edit Schedule"
-      :is-submitting="isSubmitting"
-      :visible="showEditScheduleDialog"
-      @close="showEditScheduleDialog = false"
-      @save="handleEditScheduleSave">
-      <ScheduleForm
-        ref="scheduleEditForm"
-        :audiences="audiences"
-        :disabled="isSubmitting"
-        :initial-state="editScheduleFormInitialState" />
-    </FormDialog>
-    <DeleteDialog
-      entity-type="schedule"
-      header="Delete Schedule"
-      :is-submitting="isSubmitting"
-      :visible="showDeleteScheduleDialog"
-      @close="showDeleteScheduleDialog = false"
-      @delete="handleDelete" />
+    <h4>Upcoming Schedules</h4>
+    <SchedulesGrid
+      :schedules="schedules"
+      :audiences="audiences"
+      :performers="performers" />
   </div>
 </template>
 
@@ -106,18 +31,14 @@ import { Button, Column, DataTable, useToast } from 'primevue'
 import { computed, onMounted, reactive, ref, useTemplateRef, type Ref } from 'vue'
 import DashboardSchedulesTimeslotsTable from './DashboardSchedulesTimeslotsTable.vue'
 import type { PerformerResponse } from '@/api/performers/performerResponse'
+import SchedulesGrid from './SchedulesGrid.vue'
 
 const toast = useToast()
 const isSubmitting = ref(false)
 const isLoaded = ref(false)
 const isEditingTimeslot = ref(false)
 const controlsDisabled = computed(
-  () =>
-    !isLoaded.value ||
-    isSubmitting.value ||
-    showDeleteScheduleDialog.value ||
-    showEditScheduleDialog.value ||
-    isEditingTimeslot.value
+  () => !isLoaded.value || isSubmitting.value || isEditingTimeslot.value
 )
 
 onMounted(async () => {
@@ -149,74 +70,4 @@ const handleCreateClick = async () => {
   await loadSchedules()
   createForm.value.reset()
 }
-
-const scheduleEditForm = useTemplateRef('scheduleEditForm')
-const showEditScheduleDialog = ref(false)
-let editingId = ''
-const editScheduleFormInitialState = reactive({
-  audienceId: '',
-  start: '',
-  end: '',
-  description: '',
-  startTime: new Date(),
-  endTime: new Date()
-})
-const editLimit = new Date() // Schedule times prior to this can't be altered
-setToNextHour(editLimit)
-editLimit.setDate(editLimit.getDate() - 1)
-const canEdit = (schedule: ScheduleResponse) => {
-  return Date.parse(schedule.end) > editLimit.getTime()
-}
-const handleEditScheduleActionClick = (schedule: ScheduleResponse) => {
-  editingId = schedule.id
-  editScheduleFormInitialState.audienceId = schedule.audience.id
-  editScheduleFormInitialState.start = schedule.start
-  editScheduleFormInitialState.end = schedule.end
-  editScheduleFormInitialState.description = schedule.description
-  editScheduleFormInitialState.startTime = parseDate(schedule.start)
-  editScheduleFormInitialState.endTime = parseDate(schedule.end)
-  showEditScheduleDialog.value = true
-}
-const handleEditScheduleSave = async () => {
-  if (scheduleEditForm.value == undefined) return
-  const isValid = scheduleEditForm.value.validation.validate()
-  if (!isValid) return
-
-  isSubmitting.value = true
-  const response = await api.schedules.put(editingId, scheduleEditForm.value.formState)
-  isSubmitting.value = false
-
-  if (tryHandleUnsuccessfulResponse(response, toast, scheduleEditForm.value.validation)) return
-  showEditSuccessToast(toast, 'schedule')
-  showEditScheduleDialog.value = false
-
-  const newScheduleResponse = await api.schedules.getById(editingId)
-  if (tryHandleUnsuccessfulResponse(newScheduleResponse, toast)) return
-  const indexToSplice = schedules.value.findIndex((s) => s.id == editingId)
-  schedules.value.splice(indexToSplice, 1, newScheduleResponse.data!)
-  schedules.value.sort((a, b) => {
-    if (a.start > b.start) return -1
-    if (a.start < b.start) return 1
-    return 0
-  })
-}
-
-let deletingId = ''
-const showDeleteScheduleDialog = ref(false)
-const handleDeleteScheduleActionClick = (schedule: ScheduleResponse) => {
-  deletingId = schedule.id
-  showDeleteScheduleDialog.value = true
-}
-const handleDelete = async () => {
-  showDeleteScheduleDialog.value = false
-  const response = await api.schedules.delete(deletingId)
-  tryHandleUnsuccessfulResponse(response, toast)
-  schedules.value.splice(
-    schedules.value.findIndex((s) => s.id == deletingId),
-    1
-  )
-}
-
-// TIMESLOTS
-const expandedRows = ref({})
 </script>
