+++ src/client/src/components/views/dashboard/schedules/TimeslotsGrid.vue
@@ -1,0 +1,3 @@
+<template>
+  <div class="timeslots-grid"></div>
+</template>
